Move the theme manager into WorkspaceSvg
