+++ core/workspace.js
@@ -26,8 +26,6 @@ goog.provide('Blockly.Workspace');
 goog.require('Blockly.Cursor');
 goog.require('Blockly.MarkerCursor');
 goog.require('Blockly.Events');
-goog.require('Blockly.ThemeManager');
-goog.require('Blockly.Themes.Classic');
 goog.require('Blockly.utils');
 goog.require('Blockly.utils.math');
 goog.require('Blockly.VariableMap');
@@ -129,17 +127,6 @@ Blockly.Workspace = function(opt_options) {
   this.marker_ = new Blockly.MarkerCursor();
 
   /**
-   * Object in charge of storing and updating the workspace theme.
-   * @type {!Blockly.ThemeManager}
-   * @protected
-   */
-  this.themeManager_ = this.options.parentWorkspace ?
-      this.options.parentWorkspace.getThemeManager() :
-      new Blockly.ThemeManager(this.options.theme || Blockly.Themes.Classic);
-
-  this.themeManager_.subscribeWorkspace(this);
-
-  /**
    * True if keyboard accessibility mode is on, false otherwise.
    * @type {boolean}
    * @package
@@ -208,61 +195,6 @@ Blockly.Workspace.prototype.getMarker = function() {
 };
 
 /**
- * Get the workspace theme object.
- * @return {!Blockly.Theme} The workspace theme object.
- */
-Blockly.Workspace.prototype.getTheme = function() {
-  return this.themeManager_.getTheme();
-};
-
-/**
- * Set the workspace theme object.
- * If no theme is passed, default to the `Blockly.Themes.Classic` theme.
- * @param {Blockly.Theme} theme The workspace theme object.
- */
-Blockly.Workspace.prototype.setTheme = function(theme) {
-  if (!theme) {
-    theme = /** @type {!Blockly.Theme} */ (Blockly.Themes.Classic);
-  }
-  this.themeManager_.setTheme(theme);
-};
-
-/**
- * Refresh all blocks on the workspace after a theme update.
- * @package
- */
-Blockly.Workspace.prototype.refreshTheme = function() {
-  // Update all blocks in workspace that have a style name.
-  this.updateBlockStyles_(this.getAllBlocks(false).filter(
-      function(block) {
-        return block.getStyleName() !== undefined;
-      }
-  ));
-
-  var event = new Blockly.Events.Ui(null, 'theme', null, null);
-  event.workspaceId = this.id;
-  Blockly.Events.fire(event);
-};
-
-/**
- * Updates all the blocks with new style.
- * @param {!Array.<!Blockly.Block>} blocks List of blocks to update the style
- *     on.
- * @private
- */
-Blockly.Workspace.prototype.updateBlockStyles_ = function(blocks) {
-  for (var i = 0, block; (block = blocks[i]); i++) {
-    var blockStyleName = block.getStyleName();
-    if (blockStyleName) {
-      block.setStyle(blockStyleName);
-      if (block.mutator) {
-        block.mutator.updateBlockStyle();
-      }
-    }
-  }
-};
-
-/**
  * Dispose of this workspace.
  * Unlink from all DOM elements to prevent memory leaks.
  * @suppress {checkTypes}
@@ -272,15 +204,6 @@ Blockly.Workspace.prototype.dispose = function() {
   this.clear();
   // Remove from workspace database.
   delete Blockly.Workspace.WorkspaceDB_[this.id];
-
-  if (this.themeManager_) {
-    this.themeManager_.unsubscribeWorkspace(this);
-    this.themeManager_.unsubscribe(this.svgBackground_);
-    if (!this.options.parentWorkspace) {
-      this.themeManager_.dispose();
-      this.themeManager_ = null;
-    }
-  }
 };
 
 /**
@@ -910,13 +833,4 @@ Blockly.Workspace.getAll = function() {
     workspaces.push(Blockly.Workspace.WorkspaceDB_[workspaceId]);
   }
   return workspaces;
-};
-
-/**
- * Get the theme manager for this workspace.
- * @return {!Blockly.ThemeManager} The theme manager for this workspace.
- * @package
- */
-Blockly.Workspace.prototype.getThemeManager = function() {
-  return this.themeManager_;
 };
